🚧 DGSB7B task: split maximum assimilation tests
